+++ library/ExternalContent/Cron/WpCronJobFromPostTypeSettings/WpCronJobFromPostTypeSettings.test.php
@@ -6,7 +6,7 @@ use Municipio\Config\Features\ExternalContent\ExternalContentPostTypeSettings\Ex
 use Municipio\Config\Features\ExternalContent\SourceConfig\TypesenseSourceConfigInterface;
 use Municipio\Config\Features\ExternalContent\SourceConfig\JsonSourceConfigInterface;
 use Municipio\ExternalContent\Config\SourceConfigInterface;
-use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\FilterDefinition;
+use Municipio\ExternalContent\Filter\FilterDefinition\Contracts\FilterDefinition;
 use PHPUnit\Framework\TestCase;
 use WpService\Implementations\FakeWpService;
