+++ library/ExternalContent/Filter/FilterDefinition/FilterDefinition.test.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilter\FilterDefinition;
+namespace Municipio\ExternalContent\Filter\FilterDefinition;
 
-use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\RuleSet;
+use Municipio\ExternalContent\Filter\FilterDefinition\Contracts\RuleSet;
 use PHPUnit\Framework\TestCase;
 
 class FilterDefinitionTest extends TestCase
